creating handlers for local Storage and signing transactions
